# openapi.yaml
{
  "openapi": "3.1.0",
  "info": {"title": "FactorResearch Public API", "version": "1.0.0", "description": "Stable public client contract. All operations are served below /api/v1."},
  "servers": [{"url": "/api/v1"}],
  "paths": {
    "/health": {"get": {"operationId": "getHealth", "responses": {"200": {"description": "Service is available", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HealthResponse"}}}}}}},
    "/openapi.yaml": {"get": {"operationId": "getOpenApiDocument", "responses": {"200": {"description": "This OpenAPI document", "content": {"application/yaml": {"schema": {"type": "string"}}}}}}},
    "/analysis/{symbol}": {"get": {"operationId": "getAnalysis", "parameters": [{"$ref": "#/components/parameters/Symbol"}], "responses": {
      "200": {"description": "Cached company analysis", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AnalysisResponse"}}}},
      "400": {"$ref": "#/components/responses/BadRequest"}, "404": {"$ref": "#/components/responses/NotFound"}
    }}},
    "/screener": {"get": {"operationId": "listScreenerResults", "parameters": [{"$ref": "#/components/parameters/Page"}, {"$ref": "#/components/parameters/PageSize"}], "responses": {
      "200": {"description": "Paginated screener results", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ScreenerResponse"}}}}, "400": {"$ref": "#/components/responses/BadRequest"}
    }}},
    "/account": {"get": {"operationId": "getAccount", "security": [{"bearerAuth": []}, {"sessionCookie": []}], "responses": {
      "200": {"description": "Authenticated account", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AccountResponse"}}}}, "401": {"$ref": "#/components/responses/Unauthorized"}
    }}},
    "/portfolios": {"get": {"operationId": "listPortfolios", "security": [{"bearerAuth": []}, {"sessionCookie": []}], "parameters": [{"$ref": "#/components/parameters/Page"}, {"$ref": "#/components/parameters/PageSize"}], "responses": {
      "200": {"description": "Paginated portfolio summaries", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PortfolioResponse"}}}}, "400": {"$ref": "#/components/responses/BadRequest"}, "401": {"$ref": "#/components/responses/Unauthorized"}
    }}},
    "/billing": {"get": {"operationId": "getBilling", "security": [{"bearerAuth": []}, {"sessionCookie": []}], "responses": {
      "200": {"description": "Authenticated billing status", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BillingResponse"}}}}, "401": {"$ref": "#/components/responses/Unauthorized"}
    }}}
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {"type": "http", "scheme": "bearer", "bearerFormat": "JWT"},
      "sessionCookie": {"type": "apiKey", "in": "cookie", "name": "session"}
    },
    "parameters": {
      "Symbol": {"name": "symbol", "in": "path", "required": true, "schema": {"type": "string", "pattern": "^[A-Z][A-Z0-9.-]{0,11}$"}},
      "Page": {"name": "page", "in": "query", "schema": {"type": "integer", "minimum": 1, "default": 1}},
      "PageSize": {"name": "page_size", "in": "query", "schema": {"type": "integer", "minimum": 1, "maximum": 100, "default": 25}}
    },
    "responses": {
      "BadRequest": {"description": "The request is invalid", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
      "Unauthorized": {"description": "Authentication is required", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
      "NotFound": {"description": "The resource was not found", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}}
    },
    "schemas": {
      "Meta": {"type": "object", "additionalProperties": false, "required": ["api_version", "request_id"], "properties": {"api_version": {"type": "string", "const": "v1"}, "request_id": {"type": "string"}}},
      "Pagination": {"type": "object", "additionalProperties": false, "required": ["page", "page_size", "total_items", "total_pages"], "properties": {"page": {"type": "integer"}, "page_size": {"type": "integer"}, "total_items": {"type": "integer"}, "total_pages": {"type": "integer"}}},
      "ErrorResponse": {"type": "object", "additionalProperties": false, "required": ["error", "meta"], "properties": {"error": {"type": "object", "additionalProperties": false, "required": ["code", "message"], "properties": {"code": {"type": "string"}, "message": {"type": "string"}}}, "meta": {"$ref": "#/components/schemas/Meta"}}},
      "Health": {"type": "object", "additionalProperties": false, "required": ["status"], "properties": {"status": {"type": "string", "const": "ok"}}},
      "Analysis": {"type": "object", "additionalProperties": false, "required": ["symbol", "company_name", "sector", "market_code", "price", "market_cap", "composite_score", "verdict", "analysis_version", "generated_at"], "properties": {
        "symbol": {"type": "string"}, "company_name": {"type": "string"}, "sector": {"type": "string"}, "market_code": {"type": "string"}, "price": {"type": ["number", "null"]}, "market_cap": {"type": ["number", "null"]}, "composite_score": {"type": ["number", "null"]}, "verdict": {"type": "string"}, "analysis_version": {"type": "string"}, "generated_at": {"type": ["string", "null"]}
      }},
      "ScreenerItem": {"type": "object", "additionalProperties": false, "required": ["symbol", "name", "sector", "market_code", "composite_score", "verdict", "price", "market_cap", "analyzed", "updated_at"], "properties": {
        "symbol": {"type": "string"}, "name": {"type": "string"}, "sector": {"type": "string"}, "market_code": {"type": "string"}, "composite_score": {"type": "number"}, "verdict": {"type": "string"}, "price": {"type": ["number", "null"]}, "market_cap": {"type": ["number", "null"]}, "analyzed": {"type": "boolean"}, "updated_at": {"type": ["string", "null"]}
      }},
      "PortfolioSummary": {"type": "object", "additionalProperties": false, "required": ["name", "holdings"], "properties": {"name": {"type": "string"}, "holdings": {"type": "integer", "minimum": 0}}},
      "AppearanceSettings": {"type": "object", "additionalProperties": false, "required": ["theme"], "properties": {"theme": {"type": "string", "enum": ["system", "light", "dark"]}}},
      "NotificationSettings": {"type": "object", "additionalProperties": false, "required": ["product_updates", "research_digest", "security_alerts"], "properties": {"product_updates": {"type": "boolean"}, "research_digest": {"type": "boolean"}, "security_alerts": {"type": "boolean"}}},
      "AccountSettings": {"type": "object", "additionalProperties": false, "required": ["appearance", "notifications"], "properties": {"appearance": {"$ref": "#/components/schemas/AppearanceSettings"}, "notifications": {"$ref": "#/components/schemas/NotificationSettings"}}},
      "Account": {"type": "object", "additionalProperties": false, "required": ["display_name", "auth_provider", "settings"], "properties": {"display_name": {"type": "string"}, "auth_provider": {"type": "string"}, "settings": {"$ref": "#/components/schemas/AccountSettings"}}},
      "Billing": {"type": "object", "additionalProperties": false, "required": ["plan", "status", "trial_usage", "paid"], "properties": {"plan": {"type": "string"}, "status": {"type": "string"}, "trial_usage": {"type": "integer", "minimum": 0}, "paid": {"type": "boolean"}}},
      "DataEnvelope": {"type": "object", "required": ["data", "meta"], "properties": {"data": {}, "meta": {"$ref": "#/components/schemas/Meta"}}},
      "CollectionEnvelope": {"type": "object", "required": ["data", "pagination", "meta"], "properties": {"data": {"type": "array"}, "pagination": {"$ref": "#/components/schemas/Pagination"}, "meta": {"$ref": "#/components/schemas/Meta"}}},
      "HealthResponse": {"allOf": [{"$ref": "#/components/schemas/DataEnvelope"}, {"type": "object", "properties": {"data": {"$ref": "#/components/schemas/Health"}}}]},
      "AnalysisResponse": {"allOf": [{"$ref": "#/components/schemas/DataEnvelope"}, {"type": "object", "properties": {"data": {"$ref": "#/components/schemas/Analysis"}}}]},
      "AccountResponse": {"allOf": [{"$ref": "#/components/schemas/DataEnvelope"}, {"type": "object", "properties": {"data": {"$ref": "#/components/schemas/Account"}}}]},
      "BillingResponse": {"allOf": [{"$ref": "#/components/schemas/DataEnvelope"}, {"type": "object", "properties": {"data": {"$ref": "#/components/schemas/Billing"}}}]},
      "ScreenerResponse": {"allOf": [{"$ref": "#/components/schemas/CollectionEnvelope"}, {"type": "object", "properties": {"data": {"type": "array", "items": {"$ref": "#/components/schemas/ScreenerItem"}}}}]},
      "PortfolioResponse": {"allOf": [{"$ref": "#/components/schemas/CollectionEnvelope"}, {"type": "object", "properties": {"data": {"type": "array", "items": {"$ref": "#/components/schemas/PortfolioSummary"}}}}]}
    }
  }
}